refactor(logger): improve code organization and naming consistency
Enhance the clarity and maintainability of the codebase
