+++ src/logger/index.ts
@@ -1,67 +1,18 @@
-import chalk from 'chalk'
-import durationString from './utils/duration'
-import methodString from './utils/method'
-import logString from './utils/log'
-import pathString from './utils/path'
-import statusString from './utils/status'
 import {
-  LogLevel,
   LogData,
   Logger,
-  StoreData,
+  LogLevel,
   Options,
   RequestInfo,
-  HttpError
-} from './types'
-
-/**
- * Filters log messages.
- *
- * @param {LogLevel} logLevel The log level.
- * @param {number} status The status code.
- * @param {string} method The method.
- * @param {Options} options The options.
- * @returns {boolean} `true` if the log message should be logged, otherwise `false`.
- */
-function filterLog(
-  logLevel: LogLevel,
-  status: number,
-  method: string,
-  options?: Options
-): boolean {
-  const filter = options?.config?.logFilter
-
-  if (!filter) return true
-
-  // Level
-  if (filter.level) {
-    if (Array.isArray(filter.level)) {
-      if (!filter.level.includes(logLevel)) return false
-    }
-  } else {
-    if (filter.level !== logLevel) return false
-  }
-
-  // Status
-  if (filter.status) {
-    if (Array.isArray(filter.status)) {
-      if (!filter.status.includes(status)) return false
-    }
-  } else {
-    if (filter.status !== status) return false
-  }
-
-  // Method
-  if (filter.method) {
-    if (Array.isArray(filter.method)) {
-      if (!filter.method.includes(method)) return false
-    }
-  } else {
-    if (filter.method !== method) return false
-  }
-
-  return true
-}
+  StoreData
+} from '~/types'
+import { filterLog } from './filter'
+import chalk from 'chalk'
+import logString from '~/utils/log'
+import durationString from '~/utils/duration'
+import methodString from '~/utils/method'
+import pathString from '~/utils/path'
+import statusString from '~/utils/status'
 
 /**
  * Logs a message.
@@ -138,33 +89,12 @@ function buildLogMessage(
  * @param {Options} options The options.
  * @returns {Logger} The logger.
  */
-export const createLogger = (options?: Options): Logger => ({
-  log: (level, request, data, store) =>
-    log(level, request, data, store, options),
-  customLogFormat: options?.config?.customLogFormat
-})
-
-/**
- * Handles an HTTP error.
- *
- * @param {RequestInfo} request The request.
- * @param {HttpError} error The HTTP error.
- * @param {StoreData} store The store data.
- * @param {Options} options The options.
- */
-export const handleHttpError = (
-  request: RequestInfo,
-  error: HttpError,
-  store: StoreData,
-  options?: Options
-): void => {
-  const statusCode = error.status || 500
-  const logMessage = buildLogMessage(
-    'ERROR',
-    request,
-    { status: statusCode },
-    store,
-    options
-  )
-  console.error(logMessage)
+function createLogger(options?: Options): Logger {
+  return {
+    log: (level, request, data, store) =>
+      log(level, request, data, store, options),
+    customLogFormat: options?.config?.customLogFormat
+  }
 }
+
+export { log, buildLogMessage, createLogger }
